CSCI580_Fall_2014: 3rd version for homework2. Remove z_buf from class Render (there is already a field "z" in pixel). Understand the meaning of arrange vertexes of Tri. Not only for LEE, but also for interpolate Z.

+++ CSCI580_Fall_2014/src/myGL/Pixel.java
@@ -6,13 +6,13 @@ public class Pixel
 	public short green;
 	public short blue;
 	public short alpha;
-	public int z; // signed z for clipping
+	public float z; // signed z for clipping. *****Ditong Ding***** Previously it is int, I change it to float.
 	
 	public Pixel()
 	{
 	}
 	
-	public Pixel(short red, short green, short blue, short alpha, int z)
+	public Pixel(short red, short green, short blue, short alpha, float z)
 	{
 		this.red = red;
 		this.green = green;
